GPU acceleration

+++ src/main/java/com/example/canvastest/Transform.java
@@ -4,11 +4,11 @@ public class Transform {
     // x, y =
     // ax + cy + e, bx + dy + f
 
-    public float a = 1; // Horizontal scaling
-    public float b = 0; // Horizontal skewing
-    public float c = 0; // Vertical skewing
-    public float d = 1; // Vertical scaling
+    public double a = 1; // Horizontal scaling
+    public double b = 0; // Horizontal skewing
+    public double c = 0; // Vertical skewing
+    public double d = 1; // Vertical scaling
 
-    public float e = 0; // Horizontal translation
-    public float f = 0; // Vertical translation
+    public double e = 0; // Horizontal translation
+    public double f = 0; // Vertical translation
 }
